djangochat/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-bootstrap3==24.2
django-cors-headers==4.3.1
django-filter==24.2
django-payments==2.0.0
django-phonenumber-field==7.3.0
django-rest-knox==5.0.1
djangorestframework==3.15.1
idna==3.7
mysqlclient==2.2.7
phonenumberslite==8.13.35
pillow==10.4.0
pytz==2024.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.5.0
stripe==10.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
